Add home page and template CSV
