+++ public/template.csv
@@ -1,0 +1,5 @@
+Device ID,Test Name,Result,Date,Sample ID,Operator ID
+ABL90-01,Glucose,5.6,2025-06-01,SMP001,USR001
+ABL90-02,Glucose,5.3,2025-06-01,SMP001,USR002
+ABL90-03,Glucose,6.1,2025-06-01,SMP001,USR003
+ABL90-01,Lactate,1.9,2025-06-01,SMP001,USR001
